retrace: Split off crashstats related functions
Tidies up the retrace.py a little bit.

No reason why to keep the crashstats functions in there.

A step towards issue #89
